Refactor checkpoint
