+++ handlers/get_tasks_completed.go
@@ -1,0 +1,31 @@
+// Obtener todas las tareas completadas
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func GetTasksCompleted() {
+	r.GET("/tasks/completed", func(c *gin.Context) {
+		rows, err := db.Query("SELECT id, title, description, due_date, completed FROM tasks WHERE completed=true ORDER BY id")
+		if err != nil {
+			c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
+		defer rows.Close()
+
+		tasks := []Task{}
+		for rows.Next() {
+			var task Task
+			if err := rows.Scan(&task.ID, &task.Title, &task.Description, &task.DueDate, &task.Completed); err != nil {
+				c.AbortWithError(http.StatusInternalServerError, err)
+				return
+			}
+			tasks = append(tasks, task)
+		}
+
+		c.JSON(http.StatusOK, tasks)
+	})
+}
